// types/forms.ts
export type SlideitemFormData = {
  _id?: string;
  name: string;
  slug: string;
  description: string;
  image: string;
  title?: string;
  subtitle?: string;
  btn?: string;
  textColor?: string;
  product: string;
  slide: string;
  status: "draft" | "publish" | "approve" | "reject" | "archive";
  store?: string;
  user_id?: string | null | undefined;
};