// types/schemas.ts
export const SlideitemValidationSchema: ZodType<SlideitemFormData> = z.object({
  name: z.string().regex(nameFormat, {
    message: "format: no special (&/,), min 4 and max 60 characters",
  }),
  slug: z.string().regex(slugFormat, {
    message:
      "format: no blank space (_), no special character (&,',/|...), min 4 and max 20 characters",
  }),
  description: z.string().regex(descriptionFormat, {
    message: "format:min 4 and max 500 characters",
  }),
  image: z.string().url("Image url is invalid"),
  // image: z.object({ url: z.string() }).array(),
  status: z.enum(["draft", "publish", "approve", "reject", "archive"]),
  user_id: z.string().optional().nullable(),
  slide: z.string(),
  product: z.string(),
  store: z.string().optional(),
  title: z.string().optional(),
  subtitle: z.string().optional(),
  btn: z.string().optional(),
  textColor: z.string().optional(),
});